Add localStorage to set and retrieve projects. Refactor code in modules
